Remove node mapping
