+++ src/gt4py/next/program_processors/runners/dace_fieldview/gtir_to_sdfg.py
@@ -177,11 +177,15 @@ class GtirToSDFG(eve.NodeVisitor):
             target_array = sdfg.arrays[target_node.data]
             assert not target_array.transient
             target_field_type = self._data_types[target_node.data]
-            assert isinstance(target_field_type, ts.FieldType)
 
-            subset = ",".join(
-                f"{domain_map[dim][0]}:{domain_map[dim][1]}" for dim in target_field_type.dims
-            )
+            if isinstance(target_field_type, ts.FieldType):
+                subset = ",".join(
+                    f"{domain_map[dim][0]}:{domain_map[dim][1]}" for dim in target_field_type.dims
+                )
+            else:
+                assert len(domain) == 0
+                subset = "0"
+
             dataflow_builder._head_state.add_nedge(
                 expr_node,
                 target_node,
